refactor(decimal): move component from experimental
BREAKING CHANGE: Decimal has new import path - carbon-react/lib/components/decimal
There is a codemod available to assist with this upgrade:
npx carbon-codemod move-experimental-components <target>
See https://github.com/Sage/carbon-codemod for more information.
