+++ .storybook/welcome-page/components-demo/demo/demo.component.js
@@ -3,7 +3,7 @@ import I18n from "i18n-js";
 import Button from "../../../../src/components/button";
 import Textbox from "../../../../src/components/textbox";
 import NumberInput from "../../../../src/components/number";
-import Decimal from "../../../../src/__experimental__/components/decimal";
+import Decimal from "../../../../src/components/decimal";
 import { Select, Option } from "../../../../src/components/select";
 import { Wrapper, ContentWrapper } from "../../common.style";
 import {
