+++ cypress/features/regression/experimental/decimalInput.feature
@@ -3,7 +3,7 @@ Feature: Decimal input component
 
   @positive
   Scenario Outline: Change Decimal component fieldHelp to <fieldHelp>
-    When I open default "Experimental Decimal Input Test" component in noIFrame with "decimal" json from "experimental" using "<nameOfObject>" object name
+    When I open default "Decimal Input Test" component in noIFrame with "decimal" json from "commonComponents" using "<nameOfObject>" object name
     Then fieldHelp on preview is set to <fieldHelp> in NoIFrame
     Examples:
       | fieldHelp                    | nameOfObject              |
@@ -12,7 +12,7 @@ Feature: Decimal input component
 
   @positive
   Scenario Outline: Change Decimal component label to <label>
-    When I open default "Experimental Decimal Input Test" component in noIFrame with "decimal" json from "experimental" using "<nameOfObject>" object name
+    When I open default "Decimal Input Test" component in noIFrame with "decimal" json from "commonComponents" using "<nameOfObject>" object name
     Then label on preview is <label> in NoIFrame
     Examples:
       | label                        | nameOfObject          |
@@ -21,7 +21,7 @@ Feature: Decimal input component
 
   @positive
   Scenario Outline: Change Decimal component label help to <label>
-    Given I open default "Experimental Decimal Input Test" component in noIFrame with "decimal" json from "experimental" using "<nameOfObject>" object name
+    Given I open default "Decimal Input Test" component in noIFrame with "decimal" json from "commonComponents" using "<nameOfObject>" object name
     When I hover mouse onto "question" icon in no iFrame
     Then tooltipPreview on preview is set to <label>
     Examples:
@@ -31,7 +31,7 @@ Feature: Decimal input component
 
   @positive
   Scenario Outline: Check Decimal component input field will accept characters to <input>
-    Given I open "Experimental Decimal Input Test" component page "default" in no iframe
+    Given I open "Decimal Input Test" component page "default" in no iframe
     When I set Decimal input to the "<input>"
     Then Decimal Input is set to "<input>"
     Examples:
@@ -41,6 +41,6 @@ Feature: Decimal input component
 
   @positive
   Scenario: Check Decimal component input field will accept white-space only
-    Given I open "Experimental Decimal Input Test" component page "default" in no iframe
+    Given I open "Decimal Input Test" component page "default" in no iframe
     When I set Decimal input to a string with only white-space
     Then Decimal Input is set to white-space only
