+++ cypress/features/regression/experimental/decimalPrecision.feature
@@ -3,7 +3,7 @@ Feature: Decimal component - precision
 
   @positive
   Scenario Outline: Check Decimal component input field with <nameOfObject> and input <labelInput>
-    Given I open default "Experimental Decimal Input Test" component in noIFrame with "decimal" json from "experimental" using "<nameOfObject>" object name
+    Given I open default "Decimal Input Test" component in noIFrame with "decimal" json from "commonComponents" using "<nameOfObject>" object name
     When I set Decimal input to the "<labelInput>"
     Then Decimal Input is set to "<output>"
     Examples:
