+++ cypress/fixtures/commonComponents/decimal.json
