+++ src/components/decimal/decimal.d.ts
