+++ src/components/decimal/index.d.ts
