+++ src/components/decimal/index.js
