+++ src/components/inline-inputs/inline-inputs.stories.mdx
@@ -1,70 +1,80 @@
-import { Meta, Props, Preview, Story } from '@storybook/addon-docs/blocks';
-import InlineInputs from '.';
-import Textbox from '../textbox';
-import Decimal from '../../__experimental__/components/decimal';
-import { Select, Option } from '../select';
-import { useState } from 'react';
-import { text, select } from '@storybook/addon-knobs';
-import OptionsHelper from '../../utils/helpers/options-helper';
+import { Meta, Props, Preview, Story } from "@storybook/addon-docs/blocks";
+import InlineInputs from ".";
+import Textbox from "../textbox";
+import Decimal from "../decimal";
+import { Select, Option } from "../select";
+import { useState } from "react";
+import { text, select } from "@storybook/addon-knobs";
+import OptionsHelper from "../../utils/helpers/options-helper";
 
 <Meta title="Design System/Inline Inputs" />
 
 # Inline Inputs
+
 Inline inputs allows the user to align multiple inputs.
 
 ## Contents
+
 - [Quick Start](#quick-start)
 - [Examples](#examples)
 - [Props](#props)
 
-## Quick Start 
+## Quick Start
+
 ```javascript
-import InlineInputs from 'carbon-react/lib/components/inline-inputs';
+import InlineInputs from "carbon-react/lib/components/inline-inputs";
 ```
 
-## Examples 
+## Examples
 
 ### Default
 
 <Preview>
-  <Story name="default" parameters={{ info: { disable: true }, knobs: { escapeHTML: false }}} >
-  {() => {
-    const label = text('label', 'Inline Inputs');
-    const gutter = select('gutter size', ['none', ...OptionsHelper.sizesFull], InlineInputs.defaultProps.gutter);
-    let validationProps = {};
-    validationProps = {
-      hasWarning: true,
-      inputIcon: 'warning',
-      tooltipMessage: 'warning'
-    };
-    const [decimalValue, setDecimalValue] = useState('0.00');
-    const [selectValue, setSelectValue] = useState('');
-    const handleDecimalChange = (ev) => {
-      setDecimalValue(ev.target.value.rawValue);
-    };
-    const handleSelectChange = (ev) => {
-      setSelectValue(ev.target.value);
-    };
-    return (
-      <InlineInputs label={ label } gutter={ gutter }>
-        <Textbox { ...validationProps } />
-        <Decimal value={ decimalValue } onChange={ handleDecimalChange } />
-        <Select value={ selectValue } onChange={ handleSelectChange }>
-          <Option text='Amber' value='1' />
-          <Option text='Black' value='2' />
-          <Option text='Blue' value='3' />
-          <Option text='Brown' value='4' />
-          <Option text='Green' value='5' />
-          <Option text='Orange' value='6' />
-          <Option text='Pink' value='7' />
-          <Option text='Purple' value='8' />
-          <Option text='Red' value='9' />
-          <Option text='White' value='10' />
-          <Option text='Yellow' value='11' />
-        </Select>
-      </InlineInputs>
-    )
-  }}
+  <Story
+    name="default"
+    parameters={{ info: { disable: true }, knobs: { escapeHTML: false } }}
+  >
+    {() => {
+      const label = text("label", "Inline Inputs");
+      const gutter = select(
+        "gutter size",
+        ["none", ...OptionsHelper.sizesFull],
+        InlineInputs.defaultProps.gutter
+      );
+      let validationProps = {};
+      validationProps = {
+        hasWarning: true,
+        inputIcon: "warning",
+        tooltipMessage: "warning",
+      };
+      const [decimalValue, setDecimalValue] = useState("0.00");
+      const [selectValue, setSelectValue] = useState("");
+      const handleDecimalChange = (ev) => {
+        setDecimalValue(ev.target.value.rawValue);
+      };
+      const handleSelectChange = (ev) => {
+        setSelectValue(ev.target.value);
+      };
+      return (
+        <InlineInputs label={label} gutter={gutter}>
+          <Textbox {...validationProps} />
+          <Decimal value={decimalValue} onChange={handleDecimalChange} />
+          <Select value={selectValue} onChange={handleSelectChange}>
+            <Option text="Amber" value="1" />
+            <Option text="Black" value="2" />
+            <Option text="Blue" value="3" />
+            <Option text="Brown" value="4" />
+            <Option text="Green" value="5" />
+            <Option text="Orange" value="6" />
+            <Option text="Pink" value="7" />
+            <Option text="Purple" value="8" />
+            <Option text="Red" value="9" />
+            <Option text="White" value="10" />
+            <Option text="Yellow" value="11" />
+          </Select>
+        </InlineInputs>
+      );
+    }}
   </Story>
 </Preview>
